feats: added full functionality for create page (FE-19,FE-12)
Component 'endpointForm' to re-use for the create page and the edit page, new types primarly to send the new endpoint data in the format requested. There are issues with the InputTables used for the gathering of requestVariables and responsePatterns data. The endpoint for new endpoint creation is also not declared nor used.